feat: Evaluator is now modulable
