+++ main/src/io/github/iltotore/cylang/eval/ExpressionEvaluator.scala
@@ -9,7 +9,7 @@ import scala.collection.immutable.NumericRange
 
 class ExpressionEvaluator extends Evaluator[Expression] {
 
-  override def evaluate(input: Expression)(using context: Context): EvalResult = input match {
+  override def evaluateInput(input: Expression)(using context: Context): EvalResult = input match {
 
     case Empty => Right((context, Value.Void))
 
@@ -238,7 +238,7 @@ class ExpressionEvaluator extends Evaluator[Expression] {
       if (args.length != function.parameters.length)
         abort(s"Invalid argument count. Got: ${args.length}, Expected: ${function.parameters.length}")
       val values = for (arg <- args) yield evalUnbox(arg)
-      unbox(function.evaluate(values))
+      unbox(function.evaluate(values)(using currentContext, this))
     }
 
     case ForLoop(name, from, to, step, expression) => eval {
